Fix incorrect fish path
Because this automation installs fish using homebrew, the path to the
fish shell is within homebrew, and not the usual user directories. Configuring
the default shell and registering the shell using the default path caused kitty
to not properly load fish on startup. This commit fixes that to register and
use the homebrew path.